moved util file to appropriate location
